App/serializers.py
from rest_framework import serializers
from App.models import User

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['username','first_name', 'last_name', 'contact', 'password']