requirements.txt
asgiref==3.8.0
Django==4.2
djangorestframework==3.15.0
setuptools==68.2.2
sqlparse==0.4.4
wheel==0.41.2
